Move Product pack classes in Pack subdomain in adapter
