+++ src/Adapter/Product/Pack/Repository/ProductPackRepository.php
@@ -26,7 +26,7 @@
 
 declare(strict_types=1);
 
-namespace PrestaShop\PrestaShop\Adapter\Product\Repository;
+namespace PrestaShop\PrestaShop\Adapter\Product\Pack\Repository;
 
 use Pack;
 use PrestaShop\PrestaShop\Core\Domain\Product\Combination\ValueObject\CombinationId;
